requirements.txt
Django==5.1
djangorestframework==3.15.2
django-cors-headers==4.3.1
python-dotenv==1.0.1
